Fix typings in tests.
